Make handling of a two-stage tournament almost totally correct In a two-stage tournament, the matches in the first stage refer to teams not by their participant IDs, but by their "group player IDs" (whatever that means; it's not mentioned in Challonge's API docs).
Add a `group_team_ids` field to `Team` that holds those IDs, and add a `from_id`
scope to `Team`.  `from_id` is used by code that has some form of team ID --
either a `challonge_id` or a `group_team_id` -- and it needs to get the `Team`
object corresponding to that ID.

I've tested this change with a two-stage tournament, where the first stage is
round-robin, and the second stage is single-elimination.  The tournament's
matches are shown properly, but the team records only reflect the second stage.
I'll need to update the `winner?` and `loser?` scopes in `Match` to fix that.