Remove the Match#current_match scope, use Tournament#current_match_obj instead The `Match#current_match` scope doesn't, like, work.  I think it's left over from an earlier design, but it doesn't do the right thing if you start match N, then start match M, where N > M.  It also doesn't sort its results, so it's dependent on the default sorting order of the database.
`Tournament#current_match` tracks what we think of as "the current match", so use
that instead.  Add `Tournament#current_match_obj`, which returns the `Match`
object for the current match, if there is one.  Also, when we start a match,
if there is already a current match, set its `underway_at` to nil before we
mark the new match as the current match.